% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colby.R
\name{chooseType}
\alias{chooseType}
\title{Infer what kind of color scale to draw}
\usage{
chooseType(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
one of "nonnegative", "centered", or "arbitrary"
}
\description{
Use the range of x to determine what type of color scale would work best.
}
