% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colby.R
\name{colby}
\alias{colby}
\title{Get colors from a numeric vector}
\usage{
colby(x, type = NULL, cols = NULL, quant = 0.99, log = FALSE, na_col = NA)
}
\arguments{
\item{x}{vector to color by}

\item{type}{One of "nonnegative", "centered", "arbitrary", or NULL.
\itemize{
\item If "nonnegative", the color scale will extend from 0 upwards
\item If "centered", the color scale will extend equidistant from 0 in both directions
\item If "arbitrary", the color scale will extend along the range of x
\item If NULL, then the function will do its best to choose from the above.
}}

\item{cols}{Vector of colors to form a color palette from}

\item{quant}{Data will be truncated at this quantile.}

\item{log}{If TRUE, data will be log-transformed}

\item{na_col}{Color assigned to NA elements of x}
}
\value{
A list: \enumerate{
\item col, a vector of colors corresponding to the elements of x
\item palette, the color palette used (for use in legendd)
\item legendtext, text to be used in legends
\item legendcols, colors to show in legends
}
}
\description{
Defines colors for 3 cases: non-negative vectors where position vis-a-vis 0 has meaning, vectors rounghly centered around 0, and vectors where location is irrelevant.
Tries to infer which case pertains to your vector, but this logic can be overridden with the \code{type} argument.
}
