% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colby.R
\name{drawColorLegend}
\alias{drawColorLegend}
\title{Draw a legend based on the output}
\usage{
drawColorLegend(res)
}
\arguments{
\item{res}{The list output by \code{colby}}
}
\value{
Draws a legend
}
\description{
Draws a legend based on colby() outputs in a new graphics window.
}
